Add composite template support
