+++ Source/Libs/GtkSharp/Builder.cs
@@ -32,90 +32,6 @@ namespace Gtk {
 	using System.Text;
 
 	public partial class Builder {
-
-		[System.Serializable]
-		public class HandlerNotFoundException : SystemException
-		{
-			string handler_name;
-			string signal_name;
-			System.Reflection.EventInfo evnt;
-			Type delegate_type;
-		
-			public HandlerNotFoundException (string handler_name, string signal_name,
-							 System.Reflection.EventInfo evnt, Type delegate_type)
-				: this (handler_name, signal_name, evnt, delegate_type, null)
-			{
-			}
-		
-			public HandlerNotFoundException (string handler_name, string signal_name,
-							 System.Reflection.EventInfo evnt, Type delegate_type, Exception inner)
-				: base ("No handler " + handler_name + " found for signal " + signal_name,
-					inner)
-			{
-				this.handler_name = handler_name;
-				this.signal_name = signal_name;
-				this.evnt = evnt;
-				this.delegate_type = delegate_type;
-			}
-		
-			public HandlerNotFoundException (string message, string handler_name, string signal_name,
-							 System.Reflection.EventInfo evnt, Type delegate_type)
-				: base ((message != null) ? message : "No handler " + handler_name + " found for signal " + signal_name,
-					null)
-			{
-				this.handler_name = handler_name;
-				this.signal_name = signal_name;
-				this.evnt = evnt;
-				this.delegate_type = delegate_type;
-			}
-		
-			protected HandlerNotFoundException (System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
-				: base (info, context)
-			{
-				handler_name = info.GetString ("HandlerName");
-				signal_name = info.GetString ("SignalName");
-				evnt = info.GetValue ("Event", typeof (System.Reflection.EventInfo)) as System.Reflection.EventInfo;
-				delegate_type = info.GetValue ("DelegateType", typeof (Type)) as Type;
-			}
-		
-			public string HandlerName
-			{
-				get {
-					return handler_name;
-				}
-			}
-		
-			public string SignalName
-			{
-				get {
-					return signal_name;
-				}
-			}
-		
-			public System.Reflection.EventInfo Event
-			{
-				get {
-					return evnt;
-				}
-			}
-		
-			public Type DelegateType
-			{
-				get {
-					return delegate_type;
-				}
-			}
-		
-			public override void GetObjectData (System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
-			{
-				base.GetObjectData (info, context);
-				info.AddValue ("HandlerName", handler_name);
-				info.AddValue ("SignalName", signal_name);
-				info.AddValue ("Event", evnt);
-				info.AddValue ("DelegateType", delegate_type);
-			}
-		}
-		
 		
 		[AttributeUsage (AttributeTargets.Field)]
 		public class ObjectAttribute : Attribute
@@ -194,168 +110,13 @@ namespace Gtk {
 		public void Autoconnect (object handler)
 		{
 			BindFields (handler);
-			(new SignalConnector (this, handler)).ConnectSignals ();
+			(new SignalConnector (handler)).ConnectSignals (this);
 		}
 		
 		public void Autoconnect (Type handler_class)
 		{
 			BindFields (handler_class);
-			(new SignalConnector (this, handler_class)).ConnectSignals ();
-		}
-		
-		class SignalConnector
-		{
-			Builder builder;
-			Type handler_type;
-			object handler;
-		
-			public SignalConnector (Builder builder, object handler)
-			{
-				this.builder = builder;
-				this.handler = handler;
-				handler_type = handler.GetType ();
-			}
-		
-			public SignalConnector (Builder builder, Type handler_type)
-			{
-				this.builder = builder;
-				this.handler = null;
-				this.handler_type = handler_type;		
-			}
-		
-		delegate void d_gtk_builder_connect_signals_full(IntPtr raw, GtkSharp.BuilderConnectFuncNative func, IntPtr user_data);
-		static d_gtk_builder_connect_signals_full gtk_builder_connect_signals_full = FuncLoader.LoadFunction<d_gtk_builder_connect_signals_full>(FuncLoader.GetProcAddress(GLibrary.Load(Library.Gtk), "gtk_builder_connect_signals_full"));
-		
-			public void ConnectSignals() {
-				GtkSharp.BuilderConnectFuncWrapper func_wrapper = new GtkSharp.BuilderConnectFuncWrapper (new BuilderConnectFunc (ConnectFunc));
-				gtk_builder_connect_signals_full(builder.Handle, func_wrapper.NativeDelegate, IntPtr.Zero);
-			}
-		
-			public void ConnectFunc (Builder builder, GLib.Object objekt, string signal_name, string handler_name, GLib.Object connect_object, GLib.ConnectFlags flags)
-			{
-				/* search for the event to connect */
-				System.Reflection.MemberInfo[] evnts = objekt.GetType ().
-					FindMembers (System.Reflection.MemberTypes.Event,
-					     System.Reflection.BindingFlags.Instance
-					     | System.Reflection.BindingFlags.Static
-					     | System.Reflection.BindingFlags.Public
-					     | System.Reflection.BindingFlags.NonPublic,
-					     new System.Reflection.MemberFilter (SignalFilter), signal_name);
-				foreach (System.Reflection.EventInfo ei in evnts) {
-					bool connected = false;
-					System.Reflection.MethodInfo add = ei.GetAddMethod ();
-					System.Reflection.ParameterInfo[] addpi = add.GetParameters ();
-					if (addpi.Length == 1) { /* this should be always true, unless there's something broken */
-						Type delegate_type = addpi[0].ParameterType;
-		
-						/* look for an instance method */
-						if (connect_object != null || handler != null)
-							try {
-								Delegate d = Delegate.CreateDelegate (delegate_type, connect_object != null ? connect_object : handler, handler_name);
-								add.Invoke (objekt, new object[] { d } );
-								connected = true;
-							} catch (ArgumentException) { /* ignore if there is not such instance method */
-							}
-		
-						/* look for a static method if no instance method has been found */
-						if (!connected && handler_type != null)
-							try  {
-								Delegate d = Delegate.CreateDelegate (delegate_type, handler_type, handler_name);
-								add.Invoke (objekt, new object[] { d } );
-								connected = true;
-							} catch (ArgumentException) { /* ignore if there is not such static method */
-							}
-		
-						if (!connected) {
-							string msg = ExplainError (ei.Name, delegate_type, handler_type, handler_name);
-							throw new HandlerNotFoundException (msg, handler_name, signal_name, ei, delegate_type);
-						}
-					}
-				}
-			}
-		
-			static bool SignalFilter (System.Reflection.MemberInfo m, object filterCriteria)
-			{
-				string signame = (filterCriteria as string);
-				object[] attrs = m.GetCustomAttributes (typeof (GLib.SignalAttribute), false);
-				if (attrs.Length > 0)
-				{
-					foreach (GLib.SignalAttribute a in attrs)
-					{
-						if (signame == a.CName)
-						{
-							return true;
-						}
-					}
-					return false;
-				}
-				else
-				{
-					/* this tries to match the names when no attibutes are present.
-					   It is only a fallback. */
-					signame = signame.ToLower ().Replace ("_", "");
-					string evname = m.Name.ToLower ();
-					return signame == evname;
-				}
-			}
-		
-			static string GetSignature (System.Reflection.MethodInfo method)
-			{
-				if (method == null)
-					return null;
-		
-				System.Reflection.ParameterInfo [] parameters = method.GetParameters ();
-				System.Text.StringBuilder sb = new System.Text.StringBuilder ();
-				sb.Append ('(');
-				foreach (System.Reflection.ParameterInfo info in parameters) {
-					sb.Append (info.ParameterType.ToString ());
-					sb.Append (',');
-				}
-				if (sb.Length != 0)
-					sb.Length--;
-		
-				sb.Append (')');
-				return sb.ToString ();
-			}
-		
-			static string GetSignature (Type delegate_type)
-			{
-				System.Reflection.MethodInfo method = delegate_type.GetMethod ("Invoke");
-				return GetSignature (method);
-			}
-		
-			const System.Reflection.BindingFlags flags = System.Reflection.BindingFlags.NonPublic |
-							System.Reflection.BindingFlags.Public |
-							System.Reflection.BindingFlags.Static |
-							System.Reflection.BindingFlags.Instance;
-			static string GetSignature (Type klass, string method_name)
-			{
-				try {
-					System.Reflection.MethodInfo method = klass.GetMethod (method_name, flags);
-					return GetSignature (method);
-				} catch {
-					// May be more than one method with that name and none matches
-					return null;
-				}
-			}
-		
-		
-			static string ExplainError (string event_name, Type deleg, Type klass, string method)
-			{
-				if (deleg == null || klass == null || method == null)
-					return null;
-		
-				System.Text.StringBuilder sb = new System.Text.StringBuilder ();
-				string expected = GetSignature (deleg);
-				string actual = GetSignature (klass, method);
-				if (actual == null)
-					return null;
-					sb.AppendFormat ("The handler for the event {0} should take '{1}', " +
-						"but the signature of the provided handler ('{2}') is '{3}'\n",
-						event_name, expected, method, actual);
-				return sb.ToString ();
-			}
-		
+			(new SignalConnector (handler_class)).ConnectSignals (this);
 		}
 		
 		void AddFromStream (Stream stream)
