+++ Source/Templates/GtkSharp.Template.CSharp/content/GtkSharp.CompositeWidget.CSharp/Gtk_Composite_Widget.glade
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<interface>
+  <requires lib="gtk+" version="3.18"/>
+  <template class="Gtk_Composite_Widget" parent="GtkBin">
+    <property name="visible">True</property>
+    <property name="can_focus">False</property>
+    <child>
+      <object class="GtkButton" id="Button">
+          <property name="visible">True</property>
+          <property name="can_focus">True</property>
+          <property name="receives_default">True</property>
+          <signal name="clicked" handler="button_clicked" object="Gtk_Composite_Widget" swapped="no"/>
+        </object>
+    </child>
+  </template>
+</interface>
